Fixes bug on retry of secondary rate limit error
