+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cloudnativetoolkit/git-client",
-  "version": "1.4.29",
+  "version": "1.4.30",
   "description": "",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
